pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ufz-mvgnn"
version = "0.1.0"
description = "Multi-View Graph Neural Network for Urban Functional Zone Recognition"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Yongkang Sun"},
]
keywords = ["urban computing", "graph neural network", "contrastive learning", "urban functional zone"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: GIS",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24,<3.0",
    "pandas>=2.0,<3.0",
    "geopandas>=0.14,<2.0",
    "shapely>=2.0,<3.0",
    "scipy>=1.10,<2.0",
    "torch>=2.2,<3.0",
    "torch-geometric>=2.5,<3.0",
    "scikit-learn>=1.3,<2.0",
    "umap-learn>=0.5.5,<1.0",
    "plotly>=5.18,<7.0",
    "rasterstats>=0.19,<1.0",
    "joblib>=1.3,<2.0",
    "PyYAML>=6.0,<7.0",
    "hdbscan>=0.8.33,<1.0",
    "scanpy>=1.10,<2.0",
    "anndata>=0.10,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
include = ["ufz*"]
